Mejoras del código en las sesiones de Rol y Usuario

+++ RHPsicotest.WebSite/Repositories/UserRepository.cs
@@ -5,7 +5,8 @@ using RHPsicotest.WebSite.DTOs;
 using RHPsicotest.WebSite.Models;
 using RHPsicotest.WebSite.Repositories.Contracts;
 using RHPsicotest.WebSite.Utilities;
-using RHPsicotest.WebSite.ViewModels;
+using RHPsicotest.WebSite.ViewModels.User;
+using System;
 using System.Collections.Generic;
 using System.Data;
 using System.Linq;
@@ -23,56 +24,55 @@ namespace RHPsicotest.WebSite.Repositories
             this.context = context;
         }
 
-        // Crud 
-        public Task<User> AddUser(CandidateVM user)
+
+        public async Task<bool> AddUser(UserVM userVM, List<int> rolesId)
         {
-            throw new System.NotImplementedException();
-        }
-        
-        public async Task<User> AddUser(User user, List<int> rolesId)
-        {
-            bool userExists = await UserExists(user);
+            bool result = false;
+
+            bool userExists = await UserExists(userVM);
 
             if (!userExists)
             {
-                user.Password = Helper.EncryptMD5(user.Password);
+                userVM.Password = Helper.EncryptMD5(userVM.Password);
 
-                var result = await context.Users.AddAsync(user);
-                await context.SaveChangesAsync();
+                User user = Conversion.ConvertToUser(userVM);
 
-                AddRole_User(user.IdUser, rolesId);
-
-                return result.Entity;
-            }
-
-            return null;
-        }
-        
-        public async Task<bool> UpdateUser(UserDTO userDTO, List<int> rolesId)
-        {
-            User _user = await context.Users.AsNoTracking().FirstOrDefaultAsync(u => u.Email == userDTO.Email);
-            User user = await GetUser(userDTO.IdUser);
-
-            bool result = false;
-
-            if (_user == null || _user.IdUser == userDTO.IdUser)
-            {
-                user.Name = userDTO.Name;
-                user.Email = userDTO.Email;
-
-                context.Users.Update(user);
+                await context.Users.AddAsync(user);
                 result = await context.SaveChangesAsync() > 0;
 
-                AddRole_User(user.IdUser, rolesId, true);
+                RolesAsign(user.IdUser, rolesId);
             }
 
             return result;
         }
 
-        public async Task<bool> DeleteUser(int id)
+        public async Task<bool> UpdateUser(UserUpdateVM userUpdateVM, List<int> rolesId)
         {
             bool result = false;
-            User user = await GetUser(id);
+
+            User userExist = await context.Users.AsNoTracking().FirstOrDefaultAsync(u => u.Email == userUpdateVM.Email);
+
+            if (userExist == null || userExist.IdUser == userUpdateVM.IdUser)
+            {
+                User user = await context.Users.AsNoTracking().FirstOrDefaultAsync(u => u.IdUser == userUpdateVM.IdUser);
+
+                if (userUpdateVM.Password != null) user.Password = Helper.EncryptMD5(userUpdateVM.Password);
+
+                user = Conversion.ConvertToUser(user, userUpdateVM);
+
+                context.Users.Update(user);
+                result = await context.SaveChangesAsync() > 0;
+
+                if (result) RolesAsign(user.IdUser, rolesId, true);
+            }
+
+            return result;
+        }
+
+        public async Task<bool> DeleteUser(int userId)
+        {
+            bool result = false;
+            User user = await context.Users.Include(u => u.Roles).FirstOrDefaultAsync(u => u.IdUser == userId);
 
             if (user != null)
             {
@@ -83,60 +83,40 @@ namespace RHPsicotest.WebSite.Repositories
             return result;
         }
 
-        // Get User
-        public async Task<UserDTO> GetUserDTO(int id)
+
+        // Retorna todos los usuarios en version DTO
+        // UserDTO = Datos del usuario + los roles + los permisos
+        public async Task<IEnumerable<UserDTO>> GetAllUsers()
         {
-            User user = await GetUserWithRoles(id);
+            List<User> users = await context.Users.Include(u => u.Roles).ToListAsync();
 
-            List<Role> roleList = new List<Role>();
-            List<Permission> permissionList = new List<Permission>();
-            List<Permission_Role> permission_roles = new List<Permission_Role>();
+            List<UserDTO> userDTOs = new List<UserDTO>();
 
-            foreach (var item in user.Roles)
+            foreach (User user in users)
             {
-                roleList.Add(await GetRole(item.IdRole));
+                userDTOs.Add(await GetUserDTO(user, false));
             }
 
-            foreach (var item in roleList)
-            {
-                permission_roles.AddRange(await context.Permission_Roles.Where(pr => pr.IdRole == item.IdRole).ToListAsync());
-            }
-
-            foreach (var permission in permission_roles)
-            {
-                permissionList.Add(await context.Permissions.FindAsync(permission.IdPermission));
-            }
-
-            UserDTO userDTO = Conversion.ConvertToUserDTO(user, roleList, permissionList);
-
-            return userDTO;
+            return userDTOs;
         }
-        
-        public async Task<IEnumerable<User>> GetAllUsers()
+
+
+        // Cuando el admin da clic en actualizar un usuario
+        // Retorna los datos del usuario y todos los roles.
+        // También van seleccionados los roles que tiene asignados el usuario
+        public async Task<(UserUpdateVM, MultiSelectList)> GetUserAndRolesSelected(int userId)
         {
-            return await context.Users.Include("Roles.Role").ToListAsync();
-        }
-        
-        public async Task<User> GetUserWithRoles(int id)
-        {
-            return await context.Users.Include("Roles.Role").FirstOrDefaultAsync(u => u.IdUser == id);
-        }
-        
-        public async Task<UserDTO> GetUserLogin(Login userLogin)
-        {
-            User user = await context.Users.FirstOrDefaultAsync(u => u.Email == userLogin.Email && u.Password == userLogin.Password);
+            User user = await context.Users.Include(u => u.Roles).FirstOrDefaultAsync(u => u.IdUser == userId);
 
-            UserDTO userDTO = null;
+            UserUpdateVM userUpdateVM = Conversion.ConvertToUserUpdateVM(user);
 
-            if (user != null)
-            {
-                userDTO = await GetUserDTO(user.IdUser);
-            }
+            MultiSelectList rolesList = await GetRolesSelected(user);
 
-            return userDTO;
+            return (userUpdateVM, rolesList);
         }
 
-        // Get Role
+
+        // Retorna todos los roles 
         public async Task<MultiSelectList> GetAllRoles()
         {
             List<Role> roles = await context.Roles.ToListAsync();
@@ -144,15 +124,67 @@ namespace RHPsicotest.WebSite.Repositories
             return new MultiSelectList(roles, "IdRole", "RoleName");
         }
 
-        public async Task<MultiSelectList> GetRolesSelected(int userId)
+
+
+
+        public async Task<MultiSelectList> GetRolesSelected(List<int> rolesId)
         {
-            UserDTO user = await GetUserDTO(userId);
-            IEnumerable<Role> roleUser = user.Roles;
+            List<int> selectedRoles = new List<int>();
             List<Role> roles = await context.Roles.ToListAsync();
 
-            List<int> selectedRoles = new List<int>();
+            foreach (int role in rolesId)
+            {
+                selectedRoles.Add(role);
+            }
 
-            foreach (Role role in roleUser)
+            return new MultiSelectList(roles, "IdRole", "RoleName", selectedRoles);
+        }
+
+        // ------------------------ Métodos Complementarios-----------------------------
+
+        // Convierte un usuario en DTO para mostrar los datos en una vista
+        // UserDTO = Datos del usuario + los roles + los permisos
+        private async Task<UserDTO> GetUserDTO(User user, bool withPermissions)
+        {
+            List<string> roles = new List<string>();
+            List<string> permissions = new List<string>();
+
+            List<Role> userRoles = new List<Role>();
+
+            foreach (var role in user.Roles)
+            {
+                userRoles.Add(await context.Roles.Include(r => r.Permissions).FirstOrDefaultAsync(r => r.IdRole == role.IdRole));
+            }
+
+            foreach (Role role in userRoles)
+            {
+                roles.Add(role.RoleName);
+
+                if (withPermissions)
+                {
+                    foreach (var permission in role.Permissions)
+                    {
+                        Permission permiso = await context.Permissions.FirstOrDefaultAsync(p => p.IdPermission == permission.IdPermission);
+
+                        permissions.Add(permiso.PermissionName);
+                    }
+                }
+            }
+
+            UserDTO userDTO = Conversion.ConvertToUserDTO(user, roles, permissions);
+
+            return userDTO;
+        }
+
+
+        // Método que se utiliza cuando se va actualizar un usuario
+        // Retorna todos los roles y también van seleccionados los roles del usuario
+        private async Task<MultiSelectList> GetRolesSelected(User user)
+        {
+            List<int> selectedRoles = new List<int>();
+            List<Role> roles = await context.Roles.ToListAsync();
+
+            foreach (var role in user.Roles)
             {
                 selectedRoles.Add(role.IdRole);
             }
@@ -160,32 +192,27 @@ namespace RHPsicotest.WebSite.Repositories
             return new MultiSelectList(roles, "IdRole", "RoleName", selectedRoles);
         }
 
-        // Complementary Methods
-        private async Task<User> GetUser(int id)
-        {
-            return await context.Users.FirstOrDefaultAsync(u => u.IdUser == id);
-        }
-
-        private async Task<Role> GetRole(int id)
-        {
-            return await context.Roles.FirstOrDefaultAsync(u => u.IdRole == id);
-        }
-
-        private async Task<bool> UserExists(User user)
+        private async Task<bool> UserExists(UserVM user)
         {
             return await context.Users.AnyAsync(u => u.Email == user.Email);
         }
 
-        private void AddRole_User(int userId, List<int> roles, bool delete = false)
+        // Actualiza los roles del usuario
+        private void RolesAsign(int userId, List<int> roles, bool delete = false)
         {
+            // Aquí se eliminan los roles actuales del usuario
             if (delete)
             {
                 List<Role_User> rolesUser = context.Role_Users.Where(ru => ru.IdUser == userId).ToList();
 
-                context.Role_Users.RemoveRange(rolesUser);
-                context.SaveChanges();
+                if (rolesUser != null)
+                {
+                    context.Role_Users.RemoveRange(rolesUser);
+                    context.SaveChanges();
+                }
             }
 
+            // Aquí se guardan los nuevos roles asignados
             foreach (int id in roles)
             {
                 context.Role_Users.Add(new Role_User
